Fixed bug in save_data method
save_data method in game_library would always save a new file in an
ongoing session if a file had not been loaded. Has been fixed to now
only save a new file the first time save_data is called in a session
where a file has not been loaded.